% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoxicIschemicEncephalopathy.R
\docType{package}
\name{HypoxicIschemicEncephalopathy-package}
\alias{HypoxicIschemicEncephalopathy}
\alias{HypoxicIschemicEncephalopathy-package}
\title{HypoxicIschemicEncephalopathy: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
